从 AudioEngine.jvm.kt 中剥离实现 ADB 工具的代码为独立文件
